+++ server/package.json
@@ -4,9 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "babel index.js -d dist",
+    "build": "babel ./src -d dist",
     "start": "npm run build && node dist/index.js",
-    "dev": "nodemon --exec babel-node index.js"
+    "dev": "nodemon --exec babel-node ./src/index.js"
   },
   "keywords": [],
   "author": "",
