+++ server/src/database/image/index.js
